rename files to JS convention
